// src/components/model/model.ts
export interface Employee {
  customerType: string;
  firstName: string;
  lastName: string;
  company: string;
  phoneNo: string;
  postalCode: string;
  email1: string;
  email2: string;
  email3: string;

  address1: string;
  address2: string;
  address3: string;
  address4: string;
  departmentName: string;

  remarks: string;
  gender: string;
  discountRate: string;
  doNotSDM: string;
  doNotIPO: string;
}